requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-ckeditor==6.2.0
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-formtools==2.3
django-import-export==2.6.1
django-js-asset==1.2.2
django-reversion==4.0.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
et-xmlfile==1.1.0
future==0.18.2
httplib2==0.20.2
MarkupPy==1.14
mysqlclient==2.1.0
numpy==1.21.4
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.3.4
Pillow==8.4.0
PyJWT==1.7.1
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
xadmin @ file:///Users/mac/Downloads/xadmin_bugfix-master.zip
xlrd==2.0.1
xlwt==1.3.0
